split unit tests for utils functions and set up test environment
